requirements.txt
aniso8601==9.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cors==1.0.1
filelock==3.8.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.1
future==0.18.2
gevent==22.8.0
greenlet==1.1.3.post0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
protobuf==3.20.1
PySocks==1.7.1
pytz==2022.4
requests==2.28.1
requests-file==1.5.1
six==1.16.0
SQLAlchemy==1.4.41
tldextract==3.4.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.9.0
zope.event==4.5.0
zope.interface==5.5.0

